Crate zng_state_map

Hash-map of type erased values, useful for storing assorted dynamic state.

Crate

This crate is part of the zng project.

Modules

state_map

State map helper types.

Macros

static_id

Declares a static unique ID that is lazy inited.

Structs

OwnedStateMap

Private state map.

StateId

Unique identifier of a value in a state map.

StateMapMut

Mutable state map.

StateMapRef

Read-only state map.

Traits

StateValue

Represents a type that can be a StateId value.