zng_state_map

Struct StateMapMut

pub struct StateMapMut<'a, U>(/* private fields */);

Mutable state map.

The U parameter is tag type that represents the map’s context.

Implementations

impl<'a, U> StateMapMut<'a, U>

pub fn contains<T: StateValue>(&self, id: impl Into<StateId<T>>) -> bool

Gets if the ID is set in this map.

pub fn get<T: StateValue>(&self, id: impl Into<StateId<T>>) -> Option<&T>

Reference the ID’s value set in this map.

pub fn into_get<T: StateValue>(self, id: impl Into<StateId<T>>) -> Option<&'a T>

Consume the mutable reference to the map and returns a reference to the value in the parent lifetime 'a.

pub fn copy<T: StateValue + Copy>(&self, id: impl Into<StateId<T>>) -> Option<T>

Copy the ID’s value set in this map.

pub fn get_clone<T: StateValue + Clone>( &self, id: impl Into<StateId<T>>, ) -> Option<T>

Clone the ID’s value set in this map.

pub fn req<T: StateValue>(&self, id: impl Into<StateId<T>>) -> &T

Reference the ID’s value set in this map or panics if the ID is not set.

pub fn into_req<T: StateValue>(self, id: impl Into<StateId<T>>) -> &'a T

Consume the mutable reference to the map and returns a reference to the value in the parent lifetime 'a.

pub fn flagged(&self, id: impl Into<StateId<()>>) -> bool

Gets if a state ID without value is set.

pub fn is_empty(&self) -> bool

If no state is set.

pub fn set<T: StateValue>( &mut self, id: impl Into<StateId<T>>, value: impl Into<T>, ) -> Option<T>

Set the ID’s value.

pub fn get_mut<T: StateValue>( &mut self, id: impl Into<StateId<T>>, ) -> Option<&mut T>

Mutable borrow the ID’s value set in this map.

pub fn into_get_mut<T: StateValue>( self, id: impl Into<StateId<T>>, ) -> Option<&'a mut T>

Consume the mutable reference to the map and mutable borrow the ID’s value in the parent lifetime 'a.

pub fn req_mut<T: StateValue>(&mut self, id: impl Into<StateId<T>>) -> &mut T

Mutable borrow the key value set in this map or panics if the ID is not set.

pub fn into_req_mut<T: StateValue>(self, id: impl Into<StateId<T>>) -> &'a mut T

Consume the mutable reference to the map and mutable borrow the ID value in the parent lifetime 'a.

pub fn entry<T: StateValue>( &mut self, id: impl Into<StateId<T>>, ) -> StateMapEntry<'_, T>

Gets the given ID’s corresponding entry in the map for in-place manipulation.

pub fn into_entry<T: StateValue>( self, id: impl Into<StateId<T>>, ) -> StateMapEntry<'a, T>

Consume the mutable reference to the map and returns a given ID’s corresponding entry in the map with the parent lifetime 'a.

pub fn flag(&mut self, id: impl Into<StateId<()>>) -> bool

Sets a state ID without value.

Returns if the state ID was already flagged.

pub fn reborrow(&mut self) -> StateMapMut<'_, U>

Reborrow the mutable reference.

pub fn as_ref(&self) -> StateMapRef<'_, U>

Reborrow the reference as read-only.