Module zng_state_map::state_map

State map helper types.

Structs

OccupiedStateMapEntry

A view into an occupied entry in a state map.

VacantStateMapEntry

A view into a vacant entry in a state map.

Enums

StateMapEntry

A view into a single entry in a state map, which may either be vacant or occupied.